Make base class
